RF/TEST: Move h5py compatibility hacks into a private module
