+++ nibabel/tests/test_minc2.py
@@ -13,7 +13,8 @@ from os.path import join as pjoin
 import numpy as np
 
 from .. import minc2
-from ..minc2 import Minc2File, Minc2Image, h5py, have_h5py, setup_module
+from ..minc2 import Minc2File, Minc2Image
+from .._h5py_compat import h5py, have_h5py, setup_module
 
 from nose.tools import (assert_true, assert_equal, assert_false, assert_raises)
