+++ nibabel/tests/test_proxy_api.py
@@ -46,7 +46,7 @@ from ..freesurfer.mghformat import MGHHeader
 from .. import minc1
 from ..externals.netcdf import netcdf_file
 from .. import minc2
-from ..minc2 import h5py, have_h5py
+from .._h5py_compat import h5py, have_h5py
 from .. import ecat
 from .. import parrec
